+++ docs/whats-new.rst
@@ -14,6 +14,8 @@ Coming up next (unreleased)
 Features and front-end API
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+- **New extensions `config` and `launchconfig` for** :class:`ArgoFloat` to easily access one float configuration parameter values. (:pr:`558`) by |gmaze|.
+
 - **New queries to search institution**-related entries in an :class:`ArgoIndex`: :meth:`ArgoIndex.query.institution_code`, :meth:`ArgoIndex.query.institution_name` and :meth:`ArgoIndex.query.dac`, :issue:`527`. (:pr:`528`) by |gmaze|.
 
 Breaking changes